refactor(runtime): split class/role registration walkers into named phases
ADR-0019 slice D0: a pure mechanical extraction with no behavior change,
so the D1-D9 slices can replace one phase function at a time. Source
order of every side effect is preserved; error paths keep their exact
restore/rollback (or lack thereof) semantics.

register_class_decl (one ~2,500-line function) now orchestrates named
phases:

- registration_class_validate.rs: ClassRegSnapshot (rollback snapshot
  capture/restore), check_class_role_redeclaration, class_body_is_stub,
  validate_class_parents, begin_class_def, publish_class_shell
- registration_class_compose.rs: compose_class_parent_roles (header
  does/is roles), compose_role_into_class, record_class_composed_roles,
  sharing a RoleCompositionCx/RoleCompositionOutcome instead of five
  threaded locals
- registration_class_compose_body.rs: run_composed_role_deferred_body,
  propagate_composed_role_parent_specs, install_role_puns,
  rename_generic_composed_class
- registration_class_body.rs: run_class_body (the body-walk driver and
  attribute pre-scan) plus the small statement arms; the walk state
  (name, saved package/env, class_def, own-attribute set, modifiers) is
  a ClassBodyCx passed by &mut instead of ~8 shared locals
- registration_class_body_attr.rs: class_body_has_decl (plus the
  attribute-default validation and runtime `has` registration it shares)
- registration_class_body_method.rs / _method_forms.rs:
  class_body_method_decl, the exported operator-method sub, and the
  our/my method sub-form parameter shapes (the one consolidation: the
  byte-identical our/my parameter munging became method_sub_form_params)
- registration_class_body_does.rs: class_body_does_decl (`also does`)
- registration_class_body_exit.rs: run_class_body_leave_phasers,
  persist_class_body_statics, restore_nested_type_short_names,
  finalize_class_registration, install_class_exporthow

The two statement arms that used `continue` to skip the per-statement
registry republication tail signal it via ClassBodyFlow::SkipTail so the
tail-skip behavior is unchanged.

register_role_decl (~920 lines, three passes) likewise:

- registration_role_decl.rs: check_role_body_our_scoped_decls,
  check_role_type_param_validity, role_body_is_stub,
  reset_role_registration_state, walk_role_body (pre-scan + dispatch,
  with a RoleDeclCx), finish_role_registration
- registration_role_body.rs: role_body_has_decl, role_body_does_decl
- registration_role_method.rs: role_body_method_decl

registration_class_decl.rs keeps the orchestrating entry point and
BUILTIN_PARENT_TYPES; registration_role.rs keeps the entry point and
resolve_role_candidate. All new functions are pub(super); every new file
is at or under ~500 lines.

Validated: cargo build, cargo clippy -- -D warnings, cargo fmt --check,
full `make test` (2890 files / 27469 tests PASS), full local `make
roast` (1298 files / 87164 tests PASS), and the S12/S14 targeted files.
