+++ src/runtime/mod.rs
@@ -347,8 +347,20 @@ mod registration;
 mod registration_class;
 mod registration_class_attr;
 mod registration_class_augment;
+mod registration_class_body;
+mod registration_class_body_attr;
+mod registration_class_body_does;
+mod registration_class_body_exit;
+mod registration_class_body_method;
+mod registration_class_body_method_forms;
+mod registration_class_compose;
+mod registration_class_compose_body;
 mod registration_class_decl;
+mod registration_class_validate;
 mod registration_role;
+mod registration_role_body;
+mod registration_role_decl;
+mod registration_role_method;
 pub(crate) mod registration_sub;
 mod registry;
 pub(crate) mod resolution;
