+++ src/runtime/registration_role.rs
@@ -1,10 +1,17 @@
+//! `register_role_decl` — the AST registration walker for `role`
+//! declarations — and parametric-role candidate resolution. ADR-0019 D0
+//! split the former single ~920-line walker into named pass functions with
+//! explicit inputs (see `registration_role_decl.rs`,
+//! `registration_role_body.rs`, and `registration_role_method.rs`); this
+//! file keeps the orchestrating entry point and candidate resolution.
+
 use super::registration_class::{
-    AttrValidationCtx, ResolvedRoleCandidate, builtin_role_def, looks_like_type_arg_expr,
-    make_delegation_method, parse_role_type_args, should_treat_role_arg_as_type_expr,
-    substitute_type_params_in_method, type_value_name,
+    ResolvedRoleCandidate, builtin_role_def, looks_like_type_arg_expr, parse_role_type_args,
+    should_treat_role_arg_as_type_expr,
 };
+use super::registration_role_decl::RoleDeclCx;
 use super::*;
-use crate::ast::{HandleSpec, ParamDef};
+use crate::ast::ParamDef;
 use crate::symbol::Symbol;
 
 impl Interpreter {
@@ -265,103 +272,14 @@ impl Interpreter {
     ) -> Result<(), RuntimeError> {
         self.clear_private_zeroarg_method_cache();
 
-        // Check for our-scoped declarations inside the role body.
-        // In Raku, class/subset/enum/constant/role are implicitly our-scoped,
-        // and explicit `our sub/method/variable` are also forbidden inside roles.
-        let check_body: Vec<&Stmt> = body
-            .iter()
-            .flat_map(|s| match s {
-                Stmt::SyntheticBlock(inner) => inner.iter().collect::<Vec<_>>(),
-                other => vec![other],
-            })
-            .collect();
-        for stmt in &check_body {
-            let declaration = match stmt {
-                // A `my class` inside a role is a lexically-scoped class private to the
-                // role and is allowed; only an implicitly our-scoped `class` is forbidden.
-                Stmt::ClassDecl {
-                    is_lexical: false, ..
-                } => Some("class"),
-                Stmt::ClassDecl { .. } => None,
-                // A `my subset` inside a role is lexically scoped and private to the
-                // role body, which is allowed (like `my class`/`my role`); only an
-                // implicitly our-scoped `subset` is forbidden.
-                Stmt::SubsetDecl { is_my: true, .. } => None,
-                Stmt::SubsetDecl { .. } => Some("subset"),
-                // A `my enum` is lexically scoped and private to the role body,
-                // which is allowed (like `my class`/`my subset`/`my role`); only
-                // an implicitly our-scoped `enum` is forbidden.
-                Stmt::EnumDecl { is_my: true, .. } => None,
-                Stmt::EnumDecl { .. } => Some("enum"),
-                // A `my role` is lexically scoped and private to the role body, which
-                // is allowed (like `my class`); only an implicitly our-scoped `role` is
-                // forbidden.
-                Stmt::RoleDecl { custom_traits, .. }
-                    if custom_traits.iter().any(|(t, _)| t == "__my_scoped") =>
-                {
-                    None
-                }
-                Stmt::RoleDecl { .. } => Some("role"),
-                Stmt::VarDecl {
-                    is_our: true,
-                    custom_traits,
-                    ..
-                } => {
-                    if custom_traits.iter().any(|(t, _)| t == "__constant") {
-                        Some("constant")
-                    } else {
-                        Some("variable")
-                    }
-                }
-                Stmt::SubDecl { custom_traits, .. }
-                    if custom_traits.iter().any(|(t, _)| t == "__our_scoped") =>
-                {
-                    Some("sub")
-                }
-                Stmt::MethodDecl { is_our: true, .. } => Some("method"),
-                _ => None,
-            };
-            if let Some(decl) = declaration {
-                let mut attrs = std::collections::HashMap::new();
-                attrs.insert("declaration".to_string(), Value::str(decl.to_string()));
-                attrs.insert(
-                    "message".to_string(),
-                    Value::str(format!(
-                        "Cannot declare our-scoped {} inside of a role",
-                        decl
-                    )),
-                );
-                return Err(RuntimeError::typed("X::Declaration::OurScopeInRole", attrs));
-            }
-        }
-
-        for param_def in type_param_defs {
-            if param_def.name == "__type_only__"
-                && let Some(type_name) = param_def.type_constraint.as_deref()
-                && !type_name.starts_with("::")
-                && !self.is_resolvable_type(type_name)
-            {
-                let mut attrs = std::collections::HashMap::new();
-                attrs.insert("type".to_string(), Value::str(type_name.to_string()));
-                attrs.insert(
-                    "message".to_string(),
-                    Value::str(format!(
-                        "Invalid type '{}' used in role parameter list",
-                        type_name
-                    )),
-                );
-                return Err(RuntimeError::typed("X::Parameter::InvalidType", attrs));
-            }
-        }
+        Self::check_role_body_our_scoped_decls(body)?;
+        self.check_role_type_param_validity(type_param_defs)?;
 
         // If this is a stub declaration (body is `...`, `!!!`, or `???`)
         // and a real (non-stub) role already exists under this name, treat
         // the stub as a forward declaration / no-op — don't register a new
         // stub candidate that would shadow the real one.
-        let is_stub_body = body.iter().any(|s| {
-            matches!(s, Stmt::Expr(Expr::Call { name, .. })
-                if name == "__mutsu_stub_die" || name == "__mutsu_stub_warn")
-        });
+        let is_stub_body = Self::role_body_is_stub(body);
         if is_stub_body
             && type_params.is_empty()
             && self
@@ -372,27 +290,8 @@ impl Interpreter {
         {
             return Ok(());
         }
-        // Clean up stale punned class entry for this role name.
-        self.registry_mut().classes.remove(name);
-        self.registry_mut().hidden_classes.remove(name);
-        self.registry_mut().class_composed_roles.remove(name);
-        // When registering a parametric variant of an existing non-parametric role
-        // (forming a role group), save the non-parametric role's parents so we can
-        // restore them after the parametric variant adds its own parents.
-        let prev_parents = if !type_params.is_empty()
-            && self
-                .registry()
-                .roles
-                .get(name)
-                .is_some_and(|existing| !existing.is_stub_role)
-        {
-            self.registry().role_parents.get(name).cloned()
-        } else {
-            None
-        };
-        self.registry_mut().role_parents.remove(name);
-        self.registry_mut().role_hides.remove(name);
-        let mut role_def = RoleDef {
+        let prev_parents = self.reset_role_registration_state(name, type_params);
+        let role_def = RoleDef {
             attributes: Vec::new(),
             methods: HashMap::new(),
             // A yada-body forward declaration (`role Foo { ... }`) is a stub role:
@@ -411,767 +310,25 @@ impl Interpreter {
             deferred_body_stmts: Vec::new(),
             deferred_custom_traits: Vec::new(),
         };
-        let is_parametric = !type_params.is_empty();
-        let flattened_body: Vec<&Stmt> = body
-            .iter()
-            .flat_map(|s| match s {
-                Stmt::SyntheticBlock(inner) => inner.iter().collect::<Vec<_>>(),
-                other => vec![other],
-            })
-            .collect();
-        // Pre-scan: collect attribute names declared in this role body, and the
-        // names of modules `use`d / `need`ed inside the body. A `unit role X; use
-        // A::B::C; method m(A::B::C:D $p) {...}` imports the type at BEGIN time, but
-        // this registration validates method param types before the body's `use`
-        // has loaded the module, so a qualified imported type looks unresolvable.
-        // Remember the used-module names so the param check can accept a qualified
-        // type that a body import supplies (the real resolution happens at the call
-        // site regardless).
-        let mut role_own_attrs: HashSet<String> = HashSet::new();
-        let mut body_used_modules: HashSet<String> = HashSet::new();
-        // Types declared inside the role body itself (`my enum`, `my subset`,
-        // `my class`, ...) are not yet in the registry while the role's method
-        // signatures are validated, so collect their names to accept them as
-        // parameter/attribute constraints. The real registration happens when the
-        // role body runs.
-        let mut body_declared_types: HashSet<String> = HashSet::new();
-        for stmt in &flattened_body {
-            match stmt {
-                Stmt::HasDecl {
-                    name: attr_name, ..
-                } => {
-                    role_own_attrs.insert(attr_name.resolve());
-                }
-                Stmt::Use { module, .. } | Stmt::Need { module } | Stmt::Import { module, .. } => {
-                    body_used_modules.insert(module.clone());
-                }
-                Stmt::EnumDecl { name, .. }
-                | Stmt::SubsetDecl { name, .. }
-                | Stmt::ClassDecl { name, .. }
-                | Stmt::RoleDecl { name, .. } => {
-                    body_declared_types.insert(name.resolve());
-                }
-                _ => {}
-            }
-        }
-        let role_attr_ctx = AttrValidationCtx {
-            attrs: &role_own_attrs,
-            pkg_name: name,
-            pkg_kind: "role",
+        let mut cx = RoleDeclCx {
+            name,
+            type_params,
+            role_is_rw,
+            is_parametric: !type_params.is_empty(),
+            role_def,
+            role_own_attrs: HashSet::new(),
+            body_used_modules: HashSet::new(),
+            body_declared_types: HashSet::new(),
         };
-        for stmt in flattened_body {
-            match stmt {
-                Stmt::HasDecl {
-                    name: attr_name,
-                    is_public,
-                    default,
-                    handles,
-                    is_rw,
-                    is_readonly,
-                    type_constraint,
-                    type_smiley,
-                    is_required,
-                    sigil,
-                    where_constraint,
-                    is_alias: _,
-                    is_our,
-                    is_my,
-                    is_default,
-                    is_type,
-                    deprecated_message: _,
-                    is_built: _,
-                    unknown_traits: _,
-                } => {
-                    let attr_name_str = attr_name.resolve();
-                    // A class-level (`my $.x` / `our $.x`) role attribute is NOT a
-                    // per-instance attribute: it becomes a class-level attribute on the
-                    // consuming class (accessor on the type object, `C.x`) — exactly
-                    // like `class C { my $.x }`. Record its default expr keyed by
-                    // (role, attr) and skip the per-instance attribute registration, so
-                    // no per-instance accessor shadows the class-level fallback.
-                    if *is_my || *is_our {
-                        self.registry_mut()
-                            .role_class_level_attrs
-                            .insert((name.to_string(), attr_name_str.clone()), default.clone());
-                        continue;
-                    }
-                    role_def.own_attribute_names.insert(attr_name_str.clone());
-                    // Carry an `is Type` container trait (`has @.a is G::A`) so it
-                    // can be transferred to the consuming class at composition and
-                    // its element type enforced.
-                    if let Some(it) = is_type {
-                        self.registry_mut()
-                            .role_attribute_is_types
-                            .insert((name.to_string(), attr_name_str.clone()), it.clone());
-                    }
-                    // The declared type of a role attribute (`has Int $.x`,
-                    // `has Callable %!c{Mu:U}`) is recorded per (role, attr) and
-                    // copied onto every consuming/punned class, since a role has
-                    // no class of its own to hold `attribute_types`. `::?CLASS`
-                    // stays unresolved here — it names the *consuming* class, so
-                    // it is substituted at composition.
-                    if let Some(tc) = type_constraint {
-                        self.registry_mut()
-                            .role_attribute_types
-                            .insert((name.to_string(), attr_name_str.clone()), tc.clone());
-                    }
-                    if let Some(ts) = type_smiley {
-                        self.registry_mut()
-                            .role_attribute_smileys
-                            .insert((name.to_string(), attr_name_str.clone()), ts.clone());
-                    }
-                    // `is default(...)` on a role attribute can reference the role's
-                    // type parameters (`is default(T)`), so it cannot be evaluated
-                    // until composition. Stash the expression keyed by (role, attr);
-                    // it is copied to the consuming class and evaluated at instance
-                    // construction (with type params bound).
-                    if let Some(def_expr) = is_default {
-                        self.registry_mut()
-                            .role_attribute_default_exprs
-                            .insert((name.to_string(), attr_name_str.clone()), def_expr.clone());
-                    }
-                    // Check if this attribute already exists from a composed role
-                    if let Some(existing) = role_def
-                        .attributes
-                        .iter()
-                        .find(|(n, ..)| n == &attr_name_str)
-                    {
-                        // The attribute already exists from a parent role composition.
-                        // Record the conflict; the existing one came from a composed role.
-                        // We need to figure out which role contributed it.
-                        let parent_role = self
-                            .registry()
-                            .role_parents
-                            .get(name)
-                            .and_then(|parents| {
-                                parents.iter().find(|p| {
-                                    let base =
-                                        p.split_once('[').map(|(b, _)| b).unwrap_or(p.as_str());
-                                    self.registry().roles.get(base).is_some_and(|r| {
-                                        r.attributes.iter().any(|(n, ..)| n == &attr_name_str)
-                                    })
-                                })
-                            })
-                            .cloned()
-                            .unwrap_or_else(|| "unknown".to_string());
-                        let _ = existing;
-                        role_def.attribute_conflicts.push((
-                            attr_name_str.clone(),
-                            name.to_string(),
-                            parent_role,
-                        ));
-                    }
-                    // Apply role-level `is rw`: same logic as class_is_rw
-                    // `is readonly` on individual attributes overrides `is rw` on the role
-                    let effective_is_rw = !*is_readonly && (*is_rw || (role_is_rw && *is_public));
-                    role_def.attributes.push((
-                        attr_name_str.clone(),
-                        *is_public,
-                        default.clone(),
-                        effective_is_rw,
-                        is_required.clone(),
-                        *sigil,
-                        where_constraint.as_ref().map(|wc| wc.as_ref().clone()),
-                    ));
-                    let attr_var_name = if *is_public {
-                        format!(".{}", attr_name_str)
-                    } else {
-                        format!("!{}", attr_name_str)
-                    };
-                    self.apply_handle_specs_to_role(handles, &attr_var_name, &mut role_def);
-                }
-                Stmt::DoesDecl { name: role_name } => {
-                    if *role_name == "__mutsu_role_hidden__" {
-                        role_def.is_hidden = true;
-                        continue;
-                    }
-                    let role_name_str = role_name.resolve();
-                    if let Some(hidden_name) = role_name_str.strip_prefix("__mutsu_role_hides__") {
-                        // Track hidden class relationship for this role
-                        self.registry_mut()
-                            .role_hides
-                            .entry(name.to_string())
-                            .or_default()
-                            .push(hidden_name.to_string());
-                        continue;
-                    }
-                    // A sibling role referenced by its short name (`role Derived
-                    // does Base` inside `unit module M`, where Base is registered
-                    // as `M::Base`) must resolve to its qualified name — the same
-                    // resolution the class-body DoesDecl path already does.
-                    // JSON::Unmarshal composes all its CustomUnmarshaller roles
-                    // this way.
-                    let role_name_str = if !self.registry().roles.contains_key(&role_name_str)
-                        && !self.registry().classes.contains_key(&role_name_str)
-                        && !role_name_str.contains('[')
-                    {
-                        let resolved = self.resolve_declared_type_name(&role_name_str);
-                        if self.registry().roles.contains_key(&resolved) {
-                            resolved
-                        } else {
-                            role_name_str
-                        }
-                    } else {
-                        role_name_str
-                    };
-                    if self.registry().classes.contains_key(&role_name_str) {
-                        self.registry_mut()
-                            .role_parents
-                            .entry(name.to_string())
-                            .or_default()
-                            .push(role_name_str);
-                        continue;
-                    }
-                    let base_role_name = role_name_str
-                        .split_once('[')
-                        .map(|(b, _)| b)
-                        .unwrap_or(role_name_str.as_str());
-                    if type_params.iter().any(|tp| tp == base_role_name)
-                        || (!self.registry().roles.contains_key(base_role_name)
-                            && matches!(
-                                base_role_name,
-                                "Real"
-                                    | "Numeric"
-                                    | "Cool"
-                                    | "Any"
-                                    | "Mu"
-                                    | "Positional"
-                                    | "Associative"
-                            ))
-                    {
-                        self.registry_mut()
-                            .role_parents
-                            .entry(name.to_string())
-                            .or_default()
-                            .push(role_name_str);
-                        continue;
-                    }
-                    // For a CONCRETE parametric parent (`does R1[Bool]`), resolve the
-                    // specific concretization's RoleDef by arity rather than the
-                    // by-name `roles` map — with a same-named role group (a
-                    // parameterized `R1[::T]` AND an unparameterized `R1`) the map
-                    // holds only the last-declared variant, so composing by bare name
-                    // would pull the wrong one and drop the parametric role's methods.
-                    // Skip when the application forwards a type param (`R1[::T]`),
-                    // which `resolve_role_candidate` rejects.
-                    let concretized_parent =
-                        if role_name_str.contains('[') && !role_name_str.contains("::") {
-                            self.resolve_role_candidate(&role_name_str)
-                                .ok()
-                                .flatten()
-                                .map(|(rd, _, _)| rd)
-                        } else {
-                            None
-                        };
-                    let role = match concretized_parent
-                        .or_else(|| self.registry().roles.get(base_role_name).cloned())
-                    {
-                        Some(r) => r,
-                        None => {
-                            // If trait_mod:<is> is defined and this is a lowercase name,
-                            // defer to custom trait dispatch instead of erroring.
-                            if (self.has_proto("trait_mod:<is>")
-                                || self.has_multi_candidates("trait_mod:<is>"))
-                                && role_name_str
-                                    .chars()
-                                    .next()
-                                    .is_some_and(|c| c.is_ascii_lowercase())
-                            {
-                                role_def
-                                    .deferred_custom_traits
-                                    .push(role_name_str.to_string());
-                                continue;
-                            }
-                            return Err(RuntimeError::new(format!(
-                                "Unknown role: {}",
-                                role_name_str
-                            )));
-                        }
-                    };
-                    if role.is_stub_role {
-                        return Err(RuntimeError::typed_msg(
-                            "X::Role::Parametric::NoSuchCandidate",
-                            "No matching candidate found for the parametric role",
-                        ));
-                    }
-                    self.registry_mut()
-                        .role_parents
-                        .entry(name.to_string())
-                        .or_default()
-                        .push(role_name_str.clone());
-                    // Use resolve_role_candidate to properly handle named
-                    // arguments and default values in parameterized role
-                    // composition.
-                    // `does R1[::T]` inside a parameterized role forwards this
-                    // role's type parameter into the parent role. The concrete
-                    // value of `::T` is not known until THIS role is itself
-                    // concretized, so don't resolve the parent now (that errors
-                    // with "cannot use ::T in role application"); record the
-                    // forwarding binding (parent-param -> `::T`) via the
-                    // role_type_params branch below and defer real composition to
-                    // class-application time.
-                    let forwards_type_param = role_name_str.contains("::");
-                    let type_subs: Vec<(String, String)> = if !forwards_type_param
-                        && let Some((_, resolved_param_names, resolved_values)) =
-                            self.resolve_role_candidate(&role_name_str)?
-                    {
-                        // Store the resolved param bindings so they are
-                        // available when the child role is punned to a class
-                        // and methods referencing role params are dispatched.
-                        {
-                            let mut registry = self.registry_mut();
-                            let bindings = registry
-                                .class_role_param_bindings
-                                .entry(name.to_string())
-                                .or_default();
-                            for (p, v) in resolved_param_names.iter().zip(resolved_values.iter()) {
-                                bindings.insert(p.clone(), v.clone());
-                            }
-                        }
-                        resolved_param_names
-                            .iter()
-                            .zip(resolved_values.iter())
-                            .map(|(p, v)| (p.clone(), type_value_name(v)))
-                            .collect()
-                    } else if let Some(parent_type_params) =
-                        self.registry().role_type_params.get(base_role_name)
-                    {
-                        if let Some(bracket_start) = role_name_str.find('[') {
-                            let args_str =
-                                &role_name_str[bracket_start + 1..role_name_str.len() - 1];
-                            let type_args = parse_role_type_args(args_str);
-                            parent_type_params
-                                .iter()
-                                .zip(type_args.iter())
-                                .map(|(p, a)| (p.clone(), a.clone()))
-                                .collect()
-                        } else {
-                            Vec::new()
-                        }
-                    } else {
-                        Vec::new()
-                    };
-                    for attr in &role.attributes {
-                        if role_def.attributes.iter().any(|(n, ..)| n == &attr.0) {
-                            // Already present. Only a real conflict if both
-                            // sides declared it directly (vs. inherited from
-                            // a shared ancestor in a diamond). Skip otherwise.
-                            let parent_owns = role.own_attribute_names.contains(&attr.0);
-                            let current_owns = role_def.own_attribute_names.contains(&attr.0);
-                            if parent_owns && current_owns {
-                                role_def.attribute_conflicts.push((
-                                    attr.0.clone(),
-                                    name.to_string(),
-                                    base_role_name.to_string(),
-                                ));
-                            }
-                        } else {
-                            role_def.attributes.push(attr.clone());
-                        }
-                    }
-                    for (mname, overloads) in role.methods {
-                        // Skip methods declared with `my` scope -- role-private
-                        // Submethods (is_submethod=true) ARE composed even though
-                        // they have is_my=true.
-                        let non_my_overloads: Vec<MethodDef> = overloads
-                            .into_iter()
-                            .filter(|md| !md.is_my || md.is_submethod)
-                            .collect();
-                        if non_my_overloads.is_empty() {
-                            continue;
-                        }
-                        let composed: Vec<MethodDef> = if type_subs.is_empty() {
-                            non_my_overloads
-                                .into_iter()
-                                .map(|mut md| {
-                                    if md.original_role.is_none() {
-                                        md.original_role = md.role_origin.clone();
-                                    }
-                                    md.role_origin = Some(base_role_name.to_string());
-                                    md
-                                })
-                                .collect()
-                        } else {
-                            non_my_overloads
-                                .iter()
-                                .map(|md| {
-                                    let mut method =
-                                        substitute_type_params_in_method(md, &type_subs);
-                                    if method.original_role.is_none() {
-                                        method.original_role = method.role_origin.clone();
-                                    }
-                                    method.role_origin = Some(base_role_name.to_string());
-                                    method
-                                })
-                                .collect()
-                        };
-                        role_def.methods.entry(mname).or_default().extend(composed);
-                    }
-                }
-                Stmt::MethodDecl {
-                    name: method_name,
-                    name_expr,
-                    params: _,
-                    param_defs,
-                    body: method_body,
-                    multi,
-                    is_rw,
-                    is_private,
-                    is_our: _,
-                    is_my,
-                    is_submethod,
-                    our_variable_form: _,
-                    return_type,
-                    is_default_candidate,
-                    deprecated_message: _,
-                    handles: method_handles,
-                    custom_traits: _,
-                    is_export: _,
-                    export_tags: _,
-                } => {
-                    // Validate that $!attr references in the method body are declared
-                    // in this role (same check as for class methods).
-                    Self::validate_attr_declared_in_class(&role_attr_ctx, method_body)?;
-                    // Validate that type constraints in method parameters are resolvable.
-                    // Undeclared types like A::C should throw X::Parameter::InvalidType.
-                    // A role nested in an enclosing package (e.g. `unit class A`)
-                    // registers sibling classes under the qualified name
-                    // (`A::Identifier`), but a method param may reference them by
-                    // their short name. The role's own registered name carries the
-                    // enclosing namespace (`SQL::Abstract::Renderer::SQL`); collect
-                    // every `::`-prefix of it (`SQL::Abstract::Renderer`,
-                    // `SQL::Abstract`, `SQL`) so an unqualified sibling type resolves
-                    // under whichever enclosing package actually declared it. The
-                    // role's own short name can itself be compound (`Renderer::SQL`),
-                    // so a single rsplit is not enough.
-                    let mut enclosing_prefixes: Vec<String> = Vec::new();
-                    {
-                        let mut rest = name;
-                        while let Some((pfx, _)) = rest.rsplit_once("::") {
-                            enclosing_prefixes.push(pfx.to_string());
-                            rest = pfx;
-                        }
-                    }
-                    // The role's registered short name may omit the enclosing
-                    // package (a compound role like `Renderer::SQL` is stored
-                    // without the `unit class` prefix), so also qualify with the
-                    // current package and its own `::`-prefixes.
-                    {
-                        let cur = self.current_package();
-                        let mut rest = cur.as_str();
-                        loop {
-                            if !rest.is_empty() && !enclosing_prefixes.iter().any(|p| p == rest) {
-                                enclosing_prefixes.push(rest.to_string());
-                            }
-                            match rest.rsplit_once("::") {
-                                Some((pfx, _)) => rest = pfx,
-                                None => break,
-                            }
-                        }
-                    }
-                    for pd in param_defs {
-                        if let Some(tc) = pd.type_constraint.as_deref() {
-                            // Skip type captures (::T), invocant markers, and role type params
-                            if tc.starts_with("::")
-                                || tc == "__invocant__"
-                                || type_params.iter().any(|tp| tp == tc)
-                            {
-                                continue;
-                            }
-                            // Base name of the constraint: strip definedness smiley,
-                            // coercion `(...)`, and parameterization `[...]`.
-                            let tc_base = tc
-                                .strip_suffix(":D")
-                                .or_else(|| tc.strip_suffix(":U"))
-                                .or_else(|| tc.strip_suffix(":_"))
-                                .unwrap_or(tc);
-                            let tc_base = tc_base.split(['(', '[']).next().unwrap_or(tc_base);
-                            // A role may reference its own type in a method param
-                            // (`role R { method f(R:D $x) }`); the role is not yet in
-                            // the registry while its methods validate, so accept its
-                            // own name (full or short) explicitly.
-                            let self_short = name.rsplit_once("::").map(|(_, s)| s).unwrap_or(name);
-                            let resolvable = tc_base == name
-                                || tc_base == self_short
-                                // A role type parameter carrying a definiteness
-                                // smiley (`role R[::T] { method f(T:D $x) }`,
-                                // NativeHelpers::CStruct's `LinearArray`). The
-                                // bare-`T` check above compares the whole
-                                // constraint, so only the base name matches here.
-                                || type_params.iter().any(|tp| tp == tc_base)
-                                // A type declared in this role's own body (`my enum`,
-                                // `my subset`, ...) is not registered until the body
-                                // runs; accept its name here.
-                                || body_declared_types.contains(tc_base)
-                                || self.is_resolvable_type(tc)
-                                || (!tc.contains("::")
-                                    && enclosing_prefixes.iter().any(|pfx| {
-                                        self.is_resolvable_type(&format!("{pfx}::{tc}"))
-                                    }))
-                                // Last resort: any registered type known by this
-                                // short name. A compound role name in a `unit
-                                // package` (`my role Packet::Empty`) leaves the
-                                // enclosing package out of both the role name and
-                                // `current_package`, so the prefix walk cannot
-                                // reach a sibling like `P::DecodeBuffer`; a
-                                // short-name match still finds it, mirroring how the
-                                // sub pre-pass accepts any type declared in the unit.
-                                || (!tc.contains("::")
-                                    && self.type_known_by_short_name(tc_base))
-                                // A qualified type supplied by a module `use`d within
-                                // this role body is not yet loaded at registration
-                                // time (the body's `use` runs after this validation),
-                                // so accept it if a body import could provide it. The
-                                // call site resolves it for real.
-                                || (tc.contains("::")
-                                    && body_used_modules.iter().any(|m| {
-                                        tc_base == m
-                                            || tc_base.starts_with(&format!("{m}::"))
-                                            || m.starts_with(&format!("{tc_base}::"))
-                                    }));
-                            if !resolvable {
-                                let mut attrs = std::collections::HashMap::new();
-                                attrs.insert("type".to_string(), Value::str(tc.to_string()));
-                                attrs.insert(
-                                    "message".to_string(),
-                                    Value::str(format!(
-                                        "Invalid typename '{}' in parameter declaration.",
-                                        tc
-                                    )),
-                                );
-                                return Err(RuntimeError::typed(
-                                    "X::Parameter::InvalidType",
-                                    attrs,
-                                ));
-                            }
-                        }
-                    }
-                    // Stub multi methods (body is `{...}`) that use ::?CLASS
-                    // must be implemented by the composing class.
-                    // Non-stub multi methods with ::?CLASS are fine.
-                    let body_is_stub = {
-                        let filtered: Vec<_> = method_body
-                            .iter()
-                            .filter(|s| !matches!(s, Stmt::SetLine(_)))
-                            .collect();
-                        filtered.len() == 1
-                            && matches!(
-                                filtered[0],
-                                Stmt::Expr(Expr::Call { name, .. })
-                                    if name == "__mutsu_stub_die"
-                                        || name == "__mutsu_stub_warn"
-                            )
-                    };
-                    if *multi
-                        && body_is_stub
-                        && (param_defs.iter().any(|pd| {
-                            pd.type_constraint
-                                .as_deref()
-                                .is_some_and(|tc| tc.contains("?CLASS"))
-                        }) || return_type
-                            .as_deref()
-                            .is_some_and(|rt| rt.contains("?CLASS")))
-                    {
-                        return Err(RuntimeError::typed_msg(
-                            "X::Role::Unimplemented::Multi",
-                            "Unimplemented multi method from role",
-                        ));
-                    }
-                    let resolved_method_name = if let Some(expr) = name_expr {
-                        self.eval_block_value(&[Stmt::Expr(expr.clone())])?
-                            .to_string_value()
-                    } else {
-                        method_name.resolve()
-                    };
-                    // A method always carries an implicit `*%_` slurpy so callers
-                    // can pass (or forward) named arguments the signature does not
-                    // name. Class methods get this via `effective_method_param_defs`
-                    // at registration; role methods must too, so a role-composed
-                    // method absorbs stray named args the same way a class-declared
-                    // one does.
-                    let effective_param_defs = Self::effective_method_param_defs(param_defs, false);
-                    let effective_params: Vec<String> = effective_param_defs
-                        .iter()
-                        .map(|p| p.name.clone())
-                        .collect();
-                    let def = MethodDef {
-                        lexical_package: self.current_package(),
-                        params: effective_params,
-                        param_defs: effective_param_defs,
-                        body: std::sync::Arc::new(method_body.clone()),
-                        is_rw: *is_rw,
-                        is_private: *is_private,
-                        is_multi: *multi,
-                        is_my: *is_submethod,
-                        role_origin: None,
-                        original_role: None,
-                        return_type: return_type.clone(),
-                        compiled_code: None,
-                        delegation: None,
-                        is_default: *is_default_candidate,
-                        deprecated_message: None,
-                        is_submethod: *is_submethod,
-                        captured_env: None,
-                    };
-                    // `my method` in roles are role-private, skip method table.
-                    // Submethods (is_submethod) DO get composed even though
-                    // is_my is true.
-                    let is_role_private = *is_my && !*is_submethod;
-                    if !is_role_private {
-                        if *multi {
-                            role_def
-                                .methods
-                                .entry(resolved_method_name.clone())
-                                .or_default()
-                                .push(def);
-                        } else {
-                            // A public `method STORE` and a private `method !STORE`
-                            // share the base name but are distinct methods (dispatch
-                            // filters by is_private). Preserve any existing entry of
-                            // the opposite privacy instead of overwriting it; only a
-                            // genuine same-privacy redeclaration replaces the prior
-                            // non-multi def.
-                            let entry = role_def
-                                .methods
-                                .entry(resolved_method_name.clone())
-                                .or_default();
-                            entry.retain(|d| d.is_private != def.is_private || d.is_multi);
-                            entry.push(def);
-                        }
-                    }
-                    // `handles` on a role method: synthesize forwarder methods.
-                    if !is_role_private && !method_handles.is_empty() {
-                        let source_attr_marker = format!("&{}", resolved_method_name);
-                        for spec in method_handles {
-                            match spec {
-                                HandleSpec::Name(target) => {
-                                    role_def
-                                        .methods
-                                        .entry(target.clone())
-                                        .or_default()
-                                        .push(make_delegation_method(&source_attr_marker, target));
-                                }
-                                HandleSpec::Rename { exposed, target } => {
-                                    role_def
-                                        .methods
-                                        .entry(exposed.clone())
-                                        .or_default()
-                                        .push(make_delegation_method(&source_attr_marker, target));
-                                }
-                                HandleSpec::Wildcard => {
-                                    role_def.wildcard_handles.push(source_attr_marker.clone());
-                                }
-                                HandleSpec::Regex(pattern) => {
-                                    role_def
-                                        .wildcard_handles
-                                        .push(format!("{}:regex:{}", source_attr_marker, pattern));
-                                }
-                                HandleSpec::Type(_) => {}
-                            }
-                        }
-                    }
-                }
-                Stmt::Expr(Expr::Call { name, .. })
-                    if name == "__mutsu_stub_die" || name == "__mutsu_stub_warn" =>
-                {
-                    role_def.is_stub_role = true;
-                }
-                Stmt::SetLine(_) => {
-                    // Skip source line annotations
-                }
-                _ => {
-                    if is_parametric {
-                        // Defer non-method/non-attribute statements until composition
-                        // time so they can be re-evaluated with concrete type bindings.
-                        role_def.deferred_body_stmts.push(stmt.clone());
-                    } else {
-                        // Defer execution until after the role is registered so that
-                        // role methods can be called from within the role block body
-                        // (e.g. `role R { method foo {}; R.foo }`).
-                        role_def.deferred_body_stmts.push(stmt.clone());
-                    }
-                }
-            }
-        }
-        // Capture the current environment for anonymous roles so that attribute
-        // defaults referencing closure variables can be evaluated later.
-        let has_expr_default = role_def
-            .attributes
-            .iter()
-            .any(|(_, _, default, ..)| default.is_some());
-        if has_expr_default {
-            role_def.captured_env = Some(self.env.flatten());
-        }
-        // Capture the parents that were added during this registration
-        // (these are the parents specific to this candidate).
-        let candidate_parents = self
-            .registry()
-            .role_parents
-            .get(name)
-            .cloned()
-            .unwrap_or_default();
-        let candidate = RoleCandidateDef {
-            type_params: type_params.to_vec(),
-            type_param_defs: type_param_defs.to_vec(),
-            role_def: role_def.clone(),
-            parents: candidate_parents,
-            // The revision the role was *declared* under, snapshotted at parse
-            // time (`Stmt::RoleDecl.language_version`) exactly like a class's.
-            // Reading the parser global here instead would report whatever
-            // revision happens to be active when the declaration executes,
-            // which for a role in a `use`d module is the importer's.
-            language_version: language_version.to_string(),
-        };
-        {
-            let mut registry = self.registry_mut();
-            let cands = registry
-                .role_candidates
-                .entry(name.to_string())
-                .or_default();
-            // Re-registering the same declaration (a `__hoisted` shell followed
-            // by the in-place declaration, a module body re-run, a loop) must
-            // REPLACE the same-signature candidate, not append a duplicate:
-            // `.HOW.candidates` counts these, and two same-signature candidates
-            // in one scope are impossible in Raku (roast
-            // S14-roles/parameterized-basic.t counts 3, not 6).
-            let sig_match = cands.iter().position(|c| {
-                c.type_params == candidate.type_params
-                    && format!("{:?}", c.type_param_defs)
-                        == format!("{:?}", candidate.type_param_defs)
-            });
-            match sig_match {
-                Some(i) => cands[i] = candidate,
-                None => cands.push(candidate),
-            }
-        }
-        if self
-            .registry()
-            .roles
-            .get(name)
-            .is_none_or(|existing| existing.is_stub_role || type_params.is_empty())
-        {
-            self.registry_mut().roles.insert(name.to_string(), role_def);
-            self.registry_mut()
-                .user_declared_roles
-                .insert(name.to_string());
-        }
-        if !type_params.is_empty() && !self.registry().role_type_params.contains_key(name) {
-            self.registry_mut()
-                .role_type_params
-                .insert(name.to_string(), type_params.to_vec());
-        }
-        // When a parametric variant was registered over an existing non-parametric
-        // role (forming a role group), merge the previous parents back into
-        // role_parents so that role_parent_args_for can find all candidates' parents.
-        if let Some(prev) = prev_parents {
-            let mut registry = self.registry_mut();
-            let current = registry.role_parents.entry(name.to_string()).or_default();
-            for p in prev {
-                if !current.contains(&p) {
-                    current.push(p);
-                }
-            }
-        }
+        self.walk_role_body(body, &mut cx)?;
+        self.finish_role_registration(
+            name,
+            type_params,
+            type_param_defs,
+            language_version,
+            cx.role_def,
+            prev_parents,
+        );
         Ok(())
     }
 }
